looking at curve fit prediction error for the first time fixed rem initialization pt2

+++ streamlined_mop/src/data_train.py
@@ -335,7 +335,7 @@ if __name__ == '__main__':
                 # y_an_values = [x[1][t][0] for x in error_checkpoints_an_tuples]
                 
                 #keep only the first rem elements of x_values and y_values
-                rem = np.ceil(len(x_values)/2)
+                rem = int(np.ceil(len(x_values)/2))
                 x_values_train = x_values[:rem]
                 y_values_train = y_values[:rem]
